Supports Capybara 2.4.x
Loosens the pin for this version of Kanye to support all Capybara's 2.4.x series (2.4.0-2.4.4) instead of just 2.4.4. The duck-punching should work with any of those versions, since the class being patched doesn't change in that series.

I've ticked the patch-level for Kanye to 0.0.2 here - it shouldn't affect people already using Kanye with Capybara 2.4.4, but it is a change from 0.0.1.