Remove Poltergeist version pin
I don't see any particular reason why this needs to be version-pinned. We add methods to a Poltergeist class but don't tromp all over it. This removes the pins and checks; if we run into an incompatibility down the line, we can tighten it back up.

Bumps the version because 0.0.2 was already tagged.
Conflicts:
	dmc_kanye.gemspec